♻️  zb: Split ProxyDefault into a separate mod
