+++ zbus/src/proxy/mod.rs
@@ -27,7 +27,10 @@ use crate::{
 };
 
 mod builder;
-pub use builder::{Builder, CacheProperties, ProxyDefault};
+pub use builder::{Builder, CacheProperties};
+
+mod defaults;
+pub use defaults::ProxyDefault;
 
 /// A client-side interface proxy.
 ///
